Fix draw/undo requests
Resolves #114. The "send draw offer" button will become an "accept
draw offer" button whenever a draw offer is received. The "request to
undo move" button will become an "accept undo request" button whenever
an undo request is received. The pop-ups still appear, but if they're
closed, then the user still has access to the new buttons.
